% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{mapaFolds}
\alias{mapaFolds}
\title{This function computes the one-step estimator from Jones, Gadiyar, and Vandekar (2025).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.}
\usage{
mapaFolds(y, y_hat, folds, y_hat2 = NULL, level = 0.95)
}
\arguments{
\item{y}{A numeric vector of continuous values representing the real outcome values.}

\item{y_hat}{A numeric vector of continuous values representing the predicted outcome values.}

\item{folds}{A numeric vector representing the fold number of the corresponding y_hat and y value.}

\item{y_hat2}{A numeric vector of continuous values representing the predicted outcome values from a different model.}

\item{level}{A number between 0 and 1 representing the confidence level.}
}
\value{
Returns a list with a data frame of the one-step estimator, the logit-scale one-step estimator value, and its lower bound, upper bound, and standard error; a vector of the one-step estimate values by fold for the first model; the sample number; and the confidence level. If y_hat2 is passed as an argument, the one-step estimate for y_hat2 and its difference/ratio to the one-step estimate for y_hat1 are included with confidence intervals.
}
\description{
This function computes the one-step estimator from Jones, Gadiyar, and Vandekar (2025).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
\details{
The formula for calculating the one-step estimator is:
\eqn{\hat{\rho}_{OS} = \sqrt{\text{expit}\!\left(\text{logit}\!\left({\hat{\rho}_P}^2\right)\right)+\frac{1}{n}\sum_{i=1}^{n}\left\{\frac{(Y_i -\bar{Y})^2}{\hat{\text{Var}}(\hat{\mu)}}-\frac{(Y_i - \hat{\mu}(X_{i}))^2 \, \hat{\text{Var}}(Y)}{(\hat{\text{Var}}(Y) - \hat{\text{Var}}(\hat{\mu}(X))) \,\hat{\text{Var}}(\hat{\mu}(X))}\right\}}}

The formula for calculating the one-step estimator confidence interval is:

\eqn{\sqrt{\text{expit}\!\left(\text{logit}\!\left({\hat{\rho}_{OS}}^2\right)\right)\pm z_{1 - {\alpha/2}}\sqrt{\hat{\text{Var}}\!\left\{\frac{(Y - \bar{Y})^2}{\hat{\text{Var}}(\hat{\mu})}-\frac{(Y - \hat{\mu}(X))^2 \, \hat{\text{Var}}(Y)}{(\hat{\text{Var}}(Y) - \hat{\text{Var}}(\hat{\mu}(X))) \,\hat{\text{Var}}(\hat{\mu}(X))}\right\}/ n}}}
}
