% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{sampleSplitOS}
\alias{sampleSplitOS}
\title{Perform multiple sample splits for one-step correlation estimator}
\usage{
sampleSplitOS(y, y_hat, folds, y_hat2 = NULL, level = 0.95)
}
\arguments{
\item{y}{A vector of continuous values representing the real outcome values by fold.}

\item{y_hat}{A matrix of continuous values representing the predicted outcome values by fold. Each column corresponds to a different sample split.}

\item{folds}{A matrix representing the fold number of the corresponding y_hat and y value. Each column corresponds to a different sample split.}

\item{y_hat2}{A matrix of continuous values representing the predicted outcome values by fold from a different model. Each column corresponds to a different sample split.}

\item{level}{A number between 0 and 1 representing the confidence level.}
}
\value{
Returns a list with a data frame of the one-step estimator and its bounds per split, a matrix of the one-step estimate by fold and split, the sample number, and the confidence level. If y_hat2 is passed as an argument, the one-step estimate for y_hat2 and difference between the one-step estimate for y_hat1 and y_hat2 will be provided.
}
\description{
This function computes the one-step estimator from Jones et al. (2025). Multiple sample splits are recommended for
replicability and reproducibility. The final estimate and CI bounds can be taken as the median of the results from this function.
}
\details{
The formula for calculating the one-step estimator is:
\eqn{\hat{\rho}_{OS} = \sqrt{\text{expit}\!\left(\text{logit}\!\left({\hat{\rho}_P}^2\right)\right)+\frac{1}{n}\sum_{i=1}^{n}\left\{\frac{(Y_i -\bar{Y})^2}{\hat{\text{Var}}(\hat{\mu)}}-\frac{(Y_i - \hat{\mu}(X_{i}))^2 \, \hat{\text{Var}}(Y)}{(\hat{\text{Var}}(Y) - \hat{\text{Var}}(\hat{\mu}(X))) \,\hat{\text{Var}}(\hat{\mu}(X))}\right\}}}

The formula for calculating the one-step estimator confidence interval is:

\eqn{\sqrt{\text{expit}\!\left(\text{logit}\!\left({\hat{\rho}_{OS}}^2\right)\right)\pm z_{1 - {\alpha/2}}\sqrt{\hat{\text{Var}}\!\left\{\frac{(Y - \bar{Y})^2}{\hat{\text{Var}}(\hat{\mu})}-\frac{(Y - \hat{\mu}(X))^2 \, \hat{\text{Var}}(Y)}{(\hat{\text{Var}}(Y) - \hat{\text{Var}}(\hat{\mu}(X))) \,\hat{\text{Var}}(\hat{\mu}(X))}\right\}/ n}}}
}
